+++ assets/helpers/productPrice/productPrices.js
@@ -3,9 +3,7 @@ import type { CountryGroupName, CountryGroup } from 'helpers/internationalisatio
 import { countryGroups, fromCountry } from 'helpers/internationalisation/countryGroup';
 import type { BillingPeriod } from 'helpers/billingPeriods';
 import type { FulfilmentOptions } from 'helpers/productPrice/fulfilmentOptions';
-import { NoFulfilmentOptions } from 'helpers/productPrice/fulfilmentOptions';
 import type { ProductOptions } from 'helpers/productPrice/productOptions';
-import { NoProductOptions } from 'helpers/productPrice/productOptions';
 import type { IsoCountry } from 'helpers/internationalisation/country';
 import { type IsoCurrency, glyph, extendedGlyph } from 'helpers/internationalisation/currency';
 
@@ -73,38 +71,9 @@ function applyPromotion(price: Price, promotion: ?Promotion) {
   return price;
 }
 
-function digitalPackPromotion(
-  productPrices: ProductPrices,
-  billingPeriod: BillingPeriod,
-  country: IsoCountry,
-): ?Promotion {
-  const { name } = getCountryGroup(country);
-  const currency = getCurrency(country);
-  return productPrices[name][NoFulfilmentOptions][NoProductOptions][billingPeriod][currency].promotion;
-}
-
-function digitalPackProductPrice(
-  productPrices: ProductPrices,
-  billingPeriod: BillingPeriod,
-  country: IsoCountry,
-): Price {
-  const { name } = getCountryGroup(country);
-  const currency = getCurrency(country);
-  return {
-    price: productPrices[name][NoFulfilmentOptions][NoProductOptions][billingPeriod][currency].price,
-    currency,
-  };
-}
-
-function digitalPackAmountToPay(
-  productPrices: ProductPrices,
-  billingPeriod: BillingPeriod,
-  country: IsoCountry,
-): Price {
-  const productPrice = digitalPackProductPrice(productPrices, billingPeriod, country);
-  const promotion = digitalPackPromotion(productPrices, billingPeriod, country);
-
-  return applyPromotion(productPrice, promotion);
-}
-
-export { showPrice, digitalPackPromotion, digitalPackProductPrice, digitalPackAmountToPay, applyPromotion };
+export {
+  getCurrency,
+  getCountryGroup,
+  showPrice,
+  applyPromotion,
+};
