+++ assets/pages/digital-subscription-landing/components/form.jsx
@@ -8,7 +8,8 @@ import { bindActionCreators } from 'redux';
 import { fromCountry } from 'helpers/internationalisation/countryGroup';
 import type { DigitalBillingPeriod } from 'helpers/billingPeriods';
 import { Annual, Monthly } from 'helpers/billingPeriods';
-import { digitalPackAmountToPay, showPrice, type Price, type ProductPrices } from 'helpers/productPrice/productPrices';
+import { showPrice, type Price, type ProductPrices } from 'helpers/productPrice/productPrices';
+import { amountToPay as digitalPackAmountToPay } from 'helpers/productPrice/digitalProductPrices';
 import { getDigitalPrice } from 'helpers/subscriptions';
 import { getDiscount, getFormattedFlashSalePrice, flashSaleIsActive } from 'helpers/flashSale';
 import { priceByCountryGroupId } from 'helpers/internationalisation/price';
